backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
fairlearn==0.10.0
shap==0.45.1
scipy==1.13.1
matplotlib==3.9.0
reportlab==4.1.0
openpyxl==3.1.4
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
google-generativeai==0.5.2
google-cloud-storage==2.16.0
google-cloud-bigquery==3.21.0
